+++ redash/handlers/widgets.py
@@ -4,7 +4,7 @@ from flask import request
 
 from redash import models
 from redash.wsgi import api
-from redash.permissions import require_permission
+from redash.permissions import require_permission, require_admin_or_owner
 from redash.handlers.base import BaseResource
 
 
@@ -12,12 +12,14 @@ class WidgetListAPI(BaseResource):
     @require_permission('edit_dashboard')
     def post(self):
         widget_properties = request.get_json(force=True)
+        dashboard = models.Dashboard.get_by_id_and_org(widget_properties.pop('dashboard_id'), self.current_org)
+        require_admin_or_owner(dashboard.user_id)
+
         widget_properties['options'] = json.dumps(widget_properties['options'])
         widget_properties.pop('id', None)
-        widget_properties['dashboard'] = widget_properties.pop('dashboard_id')
+        widget_properties['dashboard'] = dashboard
         widget_properties['visualization'] = widget_properties.pop('visualization_id')
-        widget = models.Widget(**widget_properties)
-        widget.save()
+        widget = models.Widget.create(**widget_properties)
 
         layout = json.loads(widget.dashboard.layout)
         new_row = True
@@ -43,10 +45,11 @@ class WidgetListAPI(BaseResource):
 class WidgetAPI(BaseResource):
     @require_permission('edit_dashboard')
     def delete(self, widget_id):
-        widget = models.Widget.get(models.Widget.id == widget_id)
+        widget = models.Widget.get_by_id_and_org(widget_id, self.current_org)
+        require_admin_or_owner(widget.dashboard.user_id)
         widget.delete_instance()
 
-        return {'layout': widget.dashboard.layout }
+        return {'layout': widget.dashboard.layout}
 
 api.add_resource(WidgetListAPI, '/api/widgets', endpoint='widgets')
 api.add_resource(WidgetAPI, '/api/widgets/<int:widget_id>', endpoint='widget')
